Backend/requirements.txt
google-generativeai
google-genai
python-dotenv
fastapi
uvicorn
python-multipart
reportlab
python-docx
pymupdf
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
pandas
openpyxl
